+++ docs/changelog.rst
@@ -5,9 +5,17 @@ ChangeLog
 
 Changes to the library are recorded here.
 
-v0.6.1
+v0.7.0
 ------
  * South-like migrations (:ref:`See Example <example_migrations>`)
+
+v0.6.2
+------
+ * Sane default __repr__ for Edges, feature parity to Vertex
+
+v0.6.1
+------
+ * Save Strategy bugfix - thanks to Nick Vollmar
 
 v0.6.0
 ------